Add context to VXC WaitFor* functions
Using a context allows for the caller to specify the timeouts it is willing to
accept, and also pass through user cancellation to better manage resources.

Corrects a couple go docs and compiler warnings. In particular, there is a
potential for Config.MakeAPICall to return a nil response causing a panic
when it makes the 'defer response.Body.Close()' call.